Rent media status updater fix

+++ frontend/src/app/pages/rent-view/rent-view.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {RentService} from "../../rent.service";
+import {MediaService} from "../../media.service";
 import {ActivatedRoute, Params, Router} from "@angular/router";
 
 @Component({
@@ -12,7 +13,10 @@ export class RentViewComponent implements OnInit {
   rent: any[];
   rentId: string;
 
-  constructor(private rentService: RentService, private route: ActivatedRoute, private router: Router) { }
+  media: any[];
+  medId:string;
+
+  constructor(private rentService: RentService, private route: ActivatedRoute, private router: Router, private mediaService: MediaService) { }
 
   ngOnInit(): void {
 
@@ -20,10 +24,15 @@ export class RentViewComponent implements OnInit {
       this.rent=rent;
     })
 
+    this.mediaService.getMedia().subscribe((media: any[]) =>{
+      this.media=media;
+    })
+
     this.route.params.subscribe(
       (params: Params) => {
         if (params.rentId) {
           this.rentId = params.rentId;
+          this.medId = params.mediaId;
         } else {
           this.rentId = undefined;
         }
@@ -32,10 +41,20 @@ export class RentViewComponent implements OnInit {
   }
 
   onClickRentDelete(){
-    this.rentService.deleteRent(this.rentId).subscribe((res: any) => {
+    if (this.rentId != undefined) {
+      this.rentService.deleteRent(this.rentId).subscribe((res: any) => {
+        console.log("Rent deleted!");
+      })
+
+      this.mediaService.updateMediaStatus(this.medId, 'AVALIABLE').subscribe((response: any) => {
+        console.log(response, "Media status set to avaliable!");
+      });
+
+      window.alert("Rent deleted!");
       this.router.navigate(['/rent']);
-      console.log();
-    })
+    } else {
+      window.alert("No rent selected!");
+    }
   }
 }
